Ensure searched collection is always converted to JSON as an array.
When converting a search query into a collection of models, if any
model is not found by its id the resulting collection will have missing
numeric keys, so it will be converted to JSON as an object instead
of as an array.

This can happen when Elasticsearch is out of sync with the application's
database, if a model has been deleted or softdeleted from the db
but is still searchable in Elasticsearch.